internal: move urlify to http module, remove dots from numbers
