+++ tests/test_plugin.py
@@ -6,7 +6,7 @@ import pytest
 from beets.library import Item
 from beets.plugins import log
 
-from beetsplug.bandcamp import DEFAULT_CONFIG, BandcampAlbumArt, BandcampPlugin, urlify
+from beetsplug.bandcamp import DEFAULT_CONFIG, BandcampAlbumArt, BandcampPlugin
 
 LABEL_URL = "https://label.bandcamp.com"
 ALBUM_URL = f"{LABEL_URL}/album/release"
@@ -62,18 +62,6 @@ def test_find_url(mb_albumid, comments, album, expected_url):
     """URLs in `mb_albumid` and `comments` fields must be found."""
     item = Item(mb_albumid=mb_albumid, comments=comments)
     assert BandcampPlugin()._find_url_in_item(item, album, "album") == expected_url
-
-
-@pytest.mark.parametrize(
-    ("title", "expected"),
-    [
-        ("LI$INGLE010 - cyberflex - LEVEL X", "li-ingle010-cyberflex-level-x"),
-        ("LI$INGLE007 - Re:drum - Movin'", "li-ingle007-re-drum-movin"),
-        ("X23 & Høbie - Exhibit A", "x23-h-bie-exhibit-a"),
-    ],
-)
-def test_urlify(title, expected):
-    assert urlify(title) == expected
 
 
 @pytest.fixture
